requirements.txt
langchain==0.2.6
langchain-community==0.2.6
langchain-google-genai
langchain-chroma==0.1.2
langchain-groq==0.1.9
google-generativeai
groq==0.9.0
chromadb==0.5.3
tiktoken==0.7.0
pypdf==4.2.0
python-dotenv==1.0.1
streamlit==1.36.0
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.0
httpx<0.28
